+++ baseten-performance-client/core/src/lib.rs
@@ -2,6 +2,7 @@ pub mod client;
 pub mod constants;
 pub mod errors;
 pub mod http;
+pub mod http_client;
 pub mod split_policy;
 pub mod utils;
 
@@ -9,5 +10,6 @@ pub use client::PerformanceClientCore;
 pub use constants::*;
 pub use errors::ClientError;
 pub use http::*;
+pub use http_client::*;
 pub use split_policy::*;
 pub use utils::*;
